+++ packages/astro/src/assets/utils/transformToPath.ts
@@ -4,7 +4,7 @@ import { shorthash } from '../../runtime/server/shorthash.js';
 import { isESMImportedImage } from '../internal.js';
 import type { ImageTransform } from '../types.js';
 
-export function propsToFilename(transform: ImageTransform) {
+export function propsToFilename(transform: ImageTransform, hash: string) {
 	if (!isESMImportedImage(transform.src)) {
 		return transform.src;
 	}
@@ -13,5 +13,12 @@ export function propsToFilename(transform: ImageTransform) {
 	const ext = extname(filename);
 	filename = basename(filename, ext);
 	const outputExt = transform.format ? `.${transform.format}` : ext;
-	return `/${filename}_${shorthash(JSON.stringify(transform))}${outputExt}`;
+	return `/${filename}_${hash}${outputExt}`;
+}
+
+export function hashTransform(transform: ImageTransform, imageService: string) {
+	// take everything from transform except alt, which is not used in the hash
+	const { alt, ...rest } = transform;
+	const hashFields = { ...rest, imageService };
+	return shorthash(JSON.stringify(hashFields));
 }