+++ packages/astro/src/content/vite-plugin-content-imports.ts
@@ -18,7 +18,6 @@ import {
 	getEntryType,
 	globalContentConfigObserver,
 	NoCollectionError,
-	patchAssets,
 	type ContentConfig,
 } from './utils.js';
 
@@ -139,7 +138,7 @@ export const _internal = {
 					});
 				}
 
-				return contentRenderer({ entry });
+				return contentRenderer.bind(this)({ entry, viteId });
 			},
 		});
 	}
@@ -232,10 +231,14 @@ export const _internal = {
 
 		const collectionConfig = contentConfig?.collections[collection];
 		let data = collectionConfig
-			? await getEntryData({ id, collection, slug, _internal, unvalidatedData }, collectionConfig)
+			? await getEntryData(
+					{ id, collection, slug, _internal, unvalidatedData },
+					collectionConfig,
+					pluginContext,
+					settings
+			  )
 			: unvalidatedData;
 
-		await patchAssets(data, pluginContext.meta.watchMode, pluginContext.emitFile, settings);
 		const contentEntryModule: ContentEntryModule = {
 			id,
 			slug,