+++ packages/astro/test/ssr-lit.test.js
@@ -25,9 +25,8 @@ describe('Lit integration in SSR', () => {
 	}
 
 	it('Is able to load', async () => {
-		delete globalThis.window; // On Windows this results in `ReferenceError: window is not defined`
 		const html = await fetchHTML('/');
 		const $ = cheerioLoad(html);
-		expect($('#win').text()).to.equal('function');
+		expect($('#str').text()).to.equal('initialized');
 	});
 });